JS Optimizations/Fixes
One place you had %75 which I’m sure you meant 75% and then I changed
all your == to === as that is more optimized. Lastly, to make your for
loops the fastest you need to do a var on the i and you want to ‘cache’
the length otherwise each time in the loop the loop will ‘ask’ what the
length of the variable is.